requirements.txt
app-version==0.2.1
awsebcli==3.7.6
blessed==1.9.5
botocore==1.4.19
cement==2.8.2
colorama==0.3.7
defusedxml==0.4.1
Django==1.11.2
django-allauth==0.25.2
django-analytical==2.0.0
django-appconf==1.0.2
django-bower==5.2.0
django-nvd3==0.9.7
django-tables2==1.2.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
Jinja2==2.8
jmespath==0.9.0
MarkupSafe==0.23
oauthlib==1.1.1
pathspec==0.3.4
psycopg2==2.6.1
python-dateutil==2.5.3
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.0.10
pytz==2017.2
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
tolerance==0.1.2
Unidecode==0.4.19
wcwidth==0.1.6
websocket-client==0.37.0
